Refactored destNav again
Made the destnav component use states to load and unload the planet components and highlight the corresponding nav button, eliminating the need for code from main.js. Implemented the code from the dest nav to work for the crew page and started implementing it in the technology page. Need to document up the code after finishing the tech page. Too tired to do it now.