Major changes to page layout
started fixing styling issues left alone for a while. Changed how the crew and destination pages are layed out. The planet and crewmember images are now in their respective page components as opposed to in their own destination planet/ crew components. This is to make creating the mobile layout easier in the future. Also working on correcting the font used as it hasn't been implemented correctly up until now. Still need to comment up the components, it will happen before the project if finished at some point.